Consolidate netstandard1.6 and netcoreapp1.1 build into .NET SDK project
There is no reason to target .netcoreapp1.1 explicitly. The netstandard1.6 target covers .NETCoreApp 1.1
