--- C5.NETStandard/C5.NETStandard.csproj
@@ -1,49 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard1.6</TargetFramework>
-    <AssemblyName>C5</AssemblyName>
-    <RootNamespace>C5</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\C5\arrays\ArrayList.cs" Link="ArrayList.cs" />
-    <Compile Include="..\C5\arrays\CircularQueue.cs" Link="CircularQueue.cs" />
-    <Compile Include="..\C5\arrays\HashedArrayList.cs" Link="HashedArrayList.cs" />
-    <Compile Include="..\C5\arrays\SortedArray.cs" Link="SortedArray.cs" />
-    <Compile Include="..\C5\Collections.cs" Link="Collections.cs" />
-    <Compile Include="..\C5\comparers\ComparerFactory.cs" Link="ComparerFactory.cs" />
-    <Compile Include="..\C5\comparers\InternalComparer.cs" Link="InternalComparer.cs" />
-    <Compile Include="..\C5\comparers\InternalEqualityComparer.cs" Link="InternalEqualityComparer.cs" />
-    <Compile Include="..\C5\Debug.cs" Link="Debug.cs" />
-    <Compile Include="..\C5\Dictionaries.cs" Link="Dictionaries.cs" />
-    <Compile Include="..\C5\EnumerableBase.cs" Link="EnumerableBase.cs" />
-    <Compile Include="..\C5\Enums.cs" Link="Enums.cs" />
-    <Compile Include="..\C5\Events.cs" Link="Events.cs" />
-    <Compile Include="..\C5\Exceptions.cs" Link="Exceptions.cs" />
-    <Compile Include="..\C5\Formatting.cs" Link="Formatting.cs" />
-    <Compile Include="..\C5\Hashers.cs" Link="Hashers.cs" />
-    <Compile Include="..\C5\hashing\HashBag.cs" Link="HashBag.cs" />
-    <Compile Include="..\C5\hashing\HashDictionary.cs" Link="HashDictionary.cs" />
-    <Compile Include="..\C5\hashing\HashTable.cs" Link="HashTable.cs" />
-    <Compile Include="..\C5\heaps\IntervalHeap.cs" Link="IntervalHeap.cs" />
-    <Compile Include="..\C5\Interfaces.cs" Link="Interfaces.cs" />
-    <Compile Include="..\C5\IReadOnlyCollection.cs" Link="IReadOnlyCollection.cs" />
-    <Compile Include="..\C5\IReadOnlyList.cs" Link="IReadOnlyList.cs" />
-    <Compile Include="..\C5\linkedlists\HashedLinkedList.cs" Link="HashedLinkedList.cs" />
-    <Compile Include="..\C5\linkedlists\LinkedList.cs" Link="LinkedList.cs" />
-    <Compile Include="..\C5\Logger.cs" Link="Logger.cs" />
-    <Compile Include="..\C5\MappedEnumerators.cs" Link="MappedEnumerators.cs" />
-    <Compile Include="..\C5\Random.cs" Link="Random.cs" />
-    <Compile Include="..\C5\Records.cs" Link="Records.cs" />
-    <Compile Include="..\C5\SerializableAttribute.cs" Link="SerializableAttribute.cs" />
-    <Compile Include="..\C5\Sorting.cs" Link="Sorting.cs" />
-    <Compile Include="..\C5\trees\RedBlackTreeBag.cs" Link="RedBlackTreeBag.cs" />
-    <Compile Include="..\C5\trees\RedBlackTreeDictionary.cs" Link="RedBlackTreeDictionary.cs" />
-    <Compile Include="..\C5\trees\RedBlackTreeSet.cs" Link="RedBlackTreeSet.cs" />
-    <Compile Include="..\C5\ViewSupport.cs" Link="ViewSupport.cs" />
-    <Compile Include="..\C5\WrappedArray.cs" Link="WrappedArray.cs" />
-    <Compile Include="..\C5\Wrappers.cs" Link="Wrappers.cs" />
-  </ItemGroup>
-
-</Project>
