--- C5.NETStandard/MemorySafeEnumerator.cs
@@ -1,107 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Threading;
-
-namespace C5
-{
-    [Serializable]
-    internal abstract class MemorySafeEnumerator<T> : IEnumerator<T>, IEnumerable<T>, IDisposable
-    {
-        //private static int MainThreadId;
-
-        //-1 means an iterator is not in use. 
-        protected int IteratorState;
-
-        protected MemoryType MemoryType { get; private set; }
-
-        protected static bool IsMainThread
-        {
-            // TODO: Update for .NET Standard 2.0 (will support Thread)...
-            get { return true; /* Thread.CurrentThread.ManagedThreadId == MainThreadId; */ }
-        }
-
-        protected MemorySafeEnumerator(MemoryType memoryType)
-        {
-            if (memoryType == MemoryType.Safe)
-            {
-                throw new NotSupportedException("Memory type safe is currently not supported in .NET Standard");
-            }
-            if (memoryType == MemoryType.Strict)
-            {
-                throw new NotSupportedException("Memory type strict is currently not supported in .NET Standard");
-            }
-            MemoryType = memoryType;
-            // TODO: Update for .NET Standard 2.0 (will support Thread)...
-            //MainThreadId = Thread.CurrentThread.ManagedThreadId;
-            IteratorState = -1;
-        }
-
-        protected abstract MemorySafeEnumerator<T> Clone();
-
-        public abstract bool MoveNext();
-
-        public abstract void Reset();
-
-        public T Current { get; protected set; }
-
-        object IEnumerator.Current
-        {
-            get { return Current; }
-        }
-
-        public virtual void Dispose()
-        {
-            IteratorState = -1;
-        }
-
-        public IEnumerator<T> GetEnumerator()
-        {
-            MemorySafeEnumerator<T> enumerator;
-
-            switch (MemoryType)
-            {
-                case MemoryType.Normal:
-                    enumerator = Clone();
-                    break;
-                case MemoryType.Safe:
-                    if (IsMainThread)
-                    {
-                        enumerator = IteratorState != -1 ? Clone() : this;
-
-                        IteratorState = 0;
-                    }
-                    else
-                    {
-                        enumerator = Clone();
-                    }
-                    break;
-                case MemoryType.Strict:
-                    if (!IsMainThread)
-                    {
-                        throw new ConcurrentEnumerationException("Multithread access detected! In Strict memory mode is not possible to iterate the collection from different threads");
-                    }
-
-                    if (IteratorState != -1)
-                    {
-                        throw new MultipleEnumerationException("Multiple Enumeration detected! In Strict memory mode is not possible to iterate the collection multiple times");
-                    }
-
-                    enumerator = this;
-                    IteratorState = 0;
-
-                    break;
-                default:
-                    throw new ArgumentOutOfRangeException();
-            }
-
-
-            return enumerator;
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-    }
-}
