+++ C5/C5.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>portable-net4+sl50+win8+wpa81+wp8;net35;net40;net45</TargetFrameworks>
+    <TargetFrameworks>portable-net4+sl50+win8+wpa81+wp8;net35;net40;net45;netstandard1.6</TargetFrameworks>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <SignAssembly>true</SignAssembly>
     <AssemblyOriginatorKeyFile>C5.snk</AssemblyOriginatorKeyFile>
